fix first track not recording
there was a bug where the first audio recording created
would not play, it seemed to fail to record altogether.
then the next track would record and play,
but the audio was the recording from the previous track.

this was due to trying to collect the chunks
from mediaRecorder too early.
so it wasn't an off-by-one error, but a race condition.

this commit adds a new state, SAVING_AUDIO, which
happens after mediaRecorder.onstop fires.
it also moves the chunk collection/audio creation
from FINISHED_RECORDING to SAVING_AUDIO.
all FINISHED_RECORDING does now is stop the mediaRecorder.