:gear: add persistent storage with dexie-react-hooks
i originally added persistent storage in 43be23a with
use-dexie, but i ran into a few problems with it.
i had to write my own types for its functions which was
time consuming, and i found the interface more cumbersome
than it was worth, considering Dexie's API is so good.
this change subclasses Dexie to add a few convenience
functions for interacting with the track table.
`useLiveQuery` works perfectly well and with the subclass
the code is v terse

i had to include a change to the tsconfig.json because
tsc was finding an error in react-dexie-hooks which
was causing the build to fail